feat(kadena-cli): move options to tx dir and add template data flag
